MapView sidebar handling fixes

+++ lib/corelib/GpsManager.cpp
@@ -45,9 +45,9 @@ bool GpsManager::iterate(uint32_t now) {
 
                 _currentPoint.lat = _gps->location.lat();
                 _currentPoint.lon = _gps->location.lng();
-                _currentPoint.alt = (float)_gps->altitude.meters();
-                speed_ = (float)_gps->speed.mps();
-                hdoop_ = (float)(_gps->hdop.isValid() ? _gps->hdop.hdop() : 99.9f);
+                _currentPoint.alt = _gps->altitude.meters();
+                speed_ = _gps->speed.mps();
+                hdoop_ = (_gps->hdop.isValid() ? _gps->hdop.hdop() : 99.9f);
                 _currentPoint.epoch = _epoch;
 
                 newFix = true;
